refactor: move date util to other folder
