+++ src/apps/ScopeChangeRequest/Components/Workflow/Criteria/Components/Criteria.tsx
@@ -4,7 +4,10 @@ import { useAtom } from '@dbeining/react-atom';
 import { WorkflowIcon } from '../../Components/WorkflowIcon';
 import { AddContributor } from './AddContributor';
 import { CriteriaStatus } from './CriteriaDetail';
-import { convertUtcToLocalDate, dateToDateTimeFormat } from '../../Utils/dateFormatting';
+import {
+    convertUtcToLocalDate,
+    dateToDateTimeFormat,
+} from '../../../../utils/helpers/dateFormatting';
 import { useScopeChangeContext } from '../../../../hooks/context/useScopeChangeContext';
 import { Contributor, Criteria } from '../../../../types/scopeChangeRequest';
 import { actionWithCommentAtom } from '../../Atoms/signingAtom';
